# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/chat/sessions/check-overdue": {
      "post": {
        "operationId": "ChatController_checkOverdueSessions_v1",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/v1/chat/sessions/{sessionId}/force-activate": {
      "post": {
        "operationId": "ChatController_forceActivateSession_v1",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/v1/chat/ably-token": {
      "get": {
        "operationId": "ChatController_getAblyToken_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/v1/chat/messages": {
      "post": {
        "operationId": "ChatController_sendMessage_v1",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/v1/chat/messages/upload": {
      "post": {
        "operationId": "ChatController_sendFileMessage_v1",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/v1/chat/history": {
      "get": {
        "operationId": "ChatController_getChatHistory_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/v1/chat/histories": {
      "get": {
        "operationId": "ChatController_getChatHistories_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/v1/chat/sessions/active": {
      "get": {
        "operationId": "ChatController_getActiveSessions_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Chat"
        ]
      }
    },
...
    "/api/v1/chat/unread-count/total": {
      "get": {
        "operationId": "ChatController_getTotalUnreadCount_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/v1/chat/sessions/{sessionId}/unread-count": {
      "get": {
        "operationId": "ChatController_getUnreadMessageCount_v1",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/v1/chat/session/{sessionId}/attachments": {
      "get": {
        "operationId": "ChatController_getAttachmentsBySessionId_v1",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/v1/chat/sessions/{sessionId}/end": {
      "put": {
        "operationId": "ChatController_endChatSession_v1",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/v1/chat/sessions/{sessionId}/enable": {
      "put": {
        "operationId": "ChatController_enableChatSession_v1",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/v1/chat/sessions/{sessionId}/status": {
      "get": {
        "operationId": "ChatController_getChatSessionStatus_v1",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/v1/chat/sessions/{sessionId}/read": {
      "put": {
        "operationId": "ChatController_markMessagesAsRead_v1",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/v1/chat/sessions/{sessionId}/presence": {
      "put": {
        "operationId": "ChatController_updateUserPresence_v1",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/v1/chat/sessions/{sessionId}/messages/read": {
      "put": {
        "operationId": "ChatController_markMessageAsRead_v1",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/v1/chat/sessions/{sessionId}/channel-history": {
      "get": {
        "operationId": "ChatController_getChannelHistory_v1",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/v1/chat/sessions/{sessionId}/upload-url": {
      "post": {
        "operationId": "ChatController_generatePresignedUploadUrl_v1",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Chat"
        ]
      }
    },
...
  },
  "info": {
    "title": "Ruang Diri API",
    "description": "API documentation for Ruang Diri application, all the services belongs to PT. Wong Makmur Sejati",
    "version": "1.0",
    "contact": {
      "name": "Support Team",
      "url": "https://ruangdiri.id",
      "email": "support@ruangdiri.id"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "termsOfService": "https://ruangdiri.id/terms"
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT token here"
      }
    },
...
  }
}